Update 07-web-monitor.html.md
This is Patty Berger Wagner.  I am proposing this change because while the language was clear, I think the example makes it more clear to people who don't know what they are doing and are just following directions.  To me, it seemed unlikely that I would just fill in one word in that key, but with the example, people could move forward with confidence.