Update 06-android-app.html.md
Patty Wagner here again.  I thought of one more place where I got tripped up and had to ask for help.  I kept hitting run and it kept running to the emulator and not to the phone.  I was able to get help on the issue from Facebook, but figure the more clear the instructions are, the fewer questions, people will have in the future.