add some nice word smithing from @jwedding + tweaks
Added a few tweaks of my own.
@jwedding did some very nice word 'smithing the bullet points.
The context he was working with was a specific letter.
I removed mentions of specific vendors to hopefully make it a little more
timeless, added a bunch of questions that might help steer towards "why" and
added some fun quote type thingies.

Also, Sarah Kaye had some good feedback on trying to get a sense of "where are
we right now" in this roadmap, still not quite sure how to adddress that.