distinguish between DIY versus nightscout as a service
Nightscout as a service better communicates the departure from DIY.  There is
such a thing as paid DIY which is the default for many spoken, silent, or
otherwise.  While I find distasteful the specific zeal for Heroku, which offers
no support to developers or community to support our concerns, the suggestion
has been to remove direct references to T1Pal.