Update index.md
Suggested edit for the front page. This documentation site should be as neutral as possible in regards to how we talk about Nightscout, with the focus being on how to get it and how to use it. Having said that, I do think it's important to set the DIY user expectations right in terms of effort needed to install and maintain their site.